requirements.txt
Flask==3.0.3
requests==2.32.3
beautifulsoup4==4.12.3
pandas==2.2.2
openpyxl==3.1.4
gunicorn==22.0.0
python-dotenv==1.0.1
psycopg2-binary
SQLAlchemy 